% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_image}
\alias{add_image}
\title{Add Image}
\usage{
add_image(imgurl, size)
}
\arguments{
\item{imgurl}{url of the image being read in by \code{magick}}

\item{size}{Size of the scaling}
}
\value{
Scaled image
}
\description{
When appending images, scale them to cut down on file size
}
\examples{
\dontrun{
see usage in `save_gifs()`
}
}
