% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_highchart}
\alias{plot_highchart}
\title{Plot Highchart}
\usage{
plot_highchart(tbl, group)
}
\arguments{
\item{tbl}{Spend data}

\item{group}{What to aggregate by - advertiser, brand, product}
}
\value{
highcharter graph
}
\description{
Function to parse spend data and plot with \code{highcharter}
}
\examples{
\dontrun{
plot_highchart(spend, advertiser)
}
}
