% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_multichart}
\alias{plot_multichart}
\title{Plot Highchart (multgroup)}
\usage{
plot_multichart(tbl, ...)
}
\arguments{
\item{tbl}{Spend data in tibble}

\item{...}{Multiple groups, either brand & product or brand & media}
}
\value{
Highchart object
}
\description{
Passing multiple arguments to group by and hardcoding the plot
}
\examples{
\dontrun{
plot_multichart(spend, brand, product)
}
}
