% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_kantar.R
\name{pull_kantar}
\alias{pull_kantar}
\title{Extract Kantar Data}
\usage{
pull_kantar(filename, asset_count)
}
\arguments{
\item{filename}{The path and file name of the Kantar data pull.}

\item{asset_count}{Number of assets per advertiser; selected via slider input in app}
}
\value{
A list with two tibbles of data; "image_out" contains the \code{magick} information for each image or gif used by an advertiser; "spend" is summarized spend data for each quarter
}
\description{
Given a specific data pull from Kantar, extract and collate spend data at various levels and also process digital imagery with the \href{https://cran.r-project.org/package=magick/vignettes/intro.html}{magick} package.
}
\examples{
\dontrun{
filename <- "~/Desktop/kantar_data_pull.xlsx"
dat <- pull_kantar(filename)
}
}
