% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_images.R
\name{render_images}
\alias{render_images}
\title{Render Images}
\usage{
render_images(tbl, size, group, n, stack_yn)
}
\arguments{
\item{tbl}{Tibble output from from \code{pull_kantar} function}

\item{size}{Size for scaling}

\item{group}{Group either a, b, or c depending on width of image}

\item{n}{Number of columns}

\item{stack_yn}{Stack images or not}
}
\value{
Images are to be renedered in a .Rmd file
}
\description{
Sort, resize and render images in order of size. If smaller width, stack in columns, otherwise stack in rows.
}
\examples{
\dontrun{
list_to_output <- split(tbl, tbl$advertiser)
render_images(list_to_output[[1]], size = 175, group = "a", n = 4, stack_yn = TRUE)
}
}
