% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_gifs.R
\name{save_gifs}
\alias{save_gifs}
\title{Save GIFs found in Kantar data}
\usage{
save_gifs(tbl)
}
\arguments{
\item{tbl}{Tibble of data returned by the \code{pull_kantar} function. Each row is a piece of creative}
}
\value{
This function saves the gifs in a new folder - "kantar_gifs" - in the current working directory. These gifs will eventually be
}
\description{
The \code{pull_kantar} uses the \link{magick} package to initially process each image. This function then looks for any GIFs in the data, reads in all images and appends them together to recreate the GIF. Because \code{magick} will run into cache issues depending on working memory, for any GIFs with more than 20 frames, the function samples every other frame to cut down the file size. Multiple gifs are appended together, and the stacking depends on the width of the gif, so that a single saved gif will contain multiple gifs that will run simultaneously.
}
\examples{
\dontrun{
save_gifs(tbl)
}
}
