% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{spend_table}
\alias{spend_table}
\title{Spend Tables}
\usage{
spend_table(tbl, choice)
}
\arguments{
\item{tbl}{Tibble spend data from \code{kantar_pull} function}

\item{choice}{Type of spend data}
}
\value{
DT datatable rendered in shiny
}
\description{
Simple function to create DT spend tables
}
\examples{
\dontrun{
spend_table(spend, brand)
}
}
