% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarise_spend}
\alias{summarise_spend}
\title{Summarise Quarterly Spend}
\usage{
summarise_spend(tbl, row)
}
\arguments{
\item{tbl}{Kantar data pull used in the \code{pull_kantar} function}

\item{row}{The row number where the data starts; determined in the \code{pull_kantar} function}
}
\value{
Tibble of summarized quarterly spend
}
\description{
Function to summarise quarterly ad spend. Can then group and summarise further as necessary.
}
\examples{
\dontrun{
summarise_spend("path-to-kantar.xlsx", row = 7)
}
}
